Refactor: fix strict type checking issues between app-schema and yjs-schema in Svelte components
Resolved multiple TypeScript strict type-checking compilation errors when running `svelte-check` that arose from incompatibilities between the `import("../../schema/app-schema").Project` and `import("../../schema/yjs-schema").Project` definitions (specifically around private fields like `childrenKeys` in the `Items` object and various property definitions in the `Item` object).

Modifications include importing both schemas to construct local union types (e.g. `AppProject | YjsProject`) where possible, and strategically falling back to `as any` alongside `eslint-disable-next-line @typescript-eslint/no-explicit-any` comments where deep structural mismatches prevented pure TypeScript assertions. This ensures Svelte type-checking compilation correctly handles either project type during Yjs data sync.
