+++ client/src/lib/yjsService.svelte.ts
@@ -208,7 +208,8 @@ export async function createNewProject(projectName: string, existingProjectId?: 
     yjsStore.yjsClient = client;
 
     if (typeof window !== "undefined") {
-        window.__CURRENT_PROJECT__ = project;
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        window.__CURRENT_PROJECT__ = project as any;
         window.__CURRENT_PROJECT_TITLE__ = projectName;
     }
