+++ client/src/routes/[project]/+layout.svelte
@@ -11,7 +11,9 @@ import { store } from "../../stores/store.svelte";
 // This layout applies to both /[project] and /[project]/[page]
 let { data, children } = $props();
 
-import type { Project } from "../../schema/app-schema";
+import type { Project as AppProject } from "../../schema/app-schema";
+import type { Project as YjsProject } from "../../schema/yjs-schema";
+type Project = AppProject | YjsProject;
 import type { YjsClient } from "../../yjs/YjsClient";
 
 let project: Project | null = $state(null);
@@ -44,7 +46,8 @@ async function loadProject(projectNameFromParam?: string) {
             yjsStore.yjsClient = client as YjsClient;
             project = client.getProject();
             // expose project to the global store so pages become available immediately
-            store.project = project;
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            store.project = project as any;
         } else {
             // Only redirect if user is authenticated (to avoid redirecting during initial load)
             // If user is not authenticated, the onMount listener will retry once auth settles.
