+++ client/src/routes/[project]/settings/+page.svelte
@@ -22,7 +22,8 @@
     } from "../../../lib/projectSnapshot";
     import type { Project } from "../../../schema/app-schema";
 
-    let project: Project | undefined = undefined;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    let project: any = undefined;
     let exportText = $state("");
     let importText = $state("");
     let importFormat = $state("opml");
@@ -63,7 +64,8 @@
                     yjsStore.yjsClient = createSnapshotClient(
                         projectName,
                         hydrated,
-                    );
+                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                    ) as any;
                 } catch {}
             }
             project = hydrated;
@@ -86,7 +88,8 @@
                 const client = await getYjsClientByProjectTitle(projectName);
                 if (client) {
                     yjsStore.yjsClient = client as unknown as import("../../../yjs/YjsClient").YjsClient;
-                    project = client.getProject();
+                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                    project = client.getProject() as any;
                 }
             } catch (err) {
                 console.warn("SettingsPage: Failed to connect to Yjs", err);
@@ -116,7 +119,8 @@
 
         if (
             (!projectForExport ||
-                projectLooksLikePlaceholder(projectForExport)) &&
+                // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                projectLooksLikePlaceholder(projectForExport as any)) &&
             projectName
         ) {
             const snapshot = loadProjectSnapshot(projectName);
@@ -135,8 +139,10 @@
         if (projectForExport) {
             exportContent =
                 format === "opml"
-                    ? exportProjectToOpml(projectForExport)
-                    : exportProjectToMarkdown(projectForExport);
+                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                    ? exportProjectToOpml(projectForExport as any)
+                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                    : exportProjectToMarkdown(projectForExport as any);
             console.log("doExport: Export content:", exportContent);
         }
 
@@ -184,10 +190,12 @@
 
         if (importFormat === "opml") {
             console.log("doImport: Calling importOpmlIntoProject");
-            importOpmlIntoProject(importText, currentProject);
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            importOpmlIntoProject(importText, currentProject as any);
         } else {
             console.log("doImport: Calling importMarkdownIntoProject");
-            importMarkdownIntoProject(importText, currentProject);
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            importMarkdownIntoProject(importText, currentProject as any);
         }
 
         console.log(
@@ -195,7 +203,8 @@
             currentProject.items?.length || 0,
         );
         if (currentProject.items && currentProject.items.length > 0) {
-            const items = currentProject.items as unknown as import("../../../schema/app-schema").Items;
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            const items = currentProject.items as any;
             const firstPage = items.at ? items.at(0) : items[0];
             const text = firstPage
                 ? typeof firstPage.text === "function"
@@ -216,7 +225,8 @@
         } else {
             // If no Yjs project was available, update as before
             console.log("doImport: No Yjs project, using fallback");
-            store.project = currentProject;
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            store.project = currentProject as any;
         }
 
         // Clear import text after successful import
@@ -236,7 +246,8 @@
 
         // Navigate to the first imported page instead of reloading
         if (currentProject.items && currentProject.items.length > 0) {
-            const items = currentProject.items as unknown as import("../../../schema/app-schema").Items;
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            const items = currentProject.items as any;
             const firstPage = items.at ? items.at(0) : items[0];
 
             if (firstPage) {
